+++ relay-server/src/actors/project.rs
@@ -75,8 +75,12 @@ pub enum Feature {
     Profiling,
     /// Enables ingestion of Session Replays (Replay Recordings and Replay Events).
     #[serde(rename = "organizations:session-replay")]
-    Replays,
-    /// Enable transaction names normalization.
+    SessionReplay,
+    /// Enables data scrubbing of replay recording payloads.
+    #[serde(rename = "organizations:session-replay-recording-scrubbing")]
+    SessionReplayRecordingScrubbing,
+    /// Enables transaction names normalization.
+    ///
     /// Replacing UUIDs, SHAs and numerical IDs by placeholders.
     #[serde(rename = "organizations:transaction-name-normalize")]
     TransactionNameNormalize,
@@ -502,6 +506,7 @@ pub struct PublicKeyConfig {
     pub numeric_id: Option<u64>,
 }
 
+#[derive(Debug)]
 struct StateChannel {
     inner: BroadcastChannel<Arc<ProjectState>>,
     no_cache: bool,
@@ -530,6 +535,7 @@ enum GetOrFetch<'a> {
 ///
 /// This structure no longer uniquely identifies a project. Instead, it identifies a project key.
 /// Projects can define multiple keys, in which case this structure is duplicated for each instance.
+#[derive(Debug)]
 pub struct Project {
     backoff: RetryBackoff,
     next_fetch_attempt: Option<Instant>,