+++ CHANGELOG.md
@@ -2,13 +2,22 @@
 
 ## Unreleased
 
-**Features**
+**Features**:
 
 - Add support for client hints. ([#1752](https://github.com/getsentry/relay/pull/1752))
+- Apply all configured data scrubbing rules on Replays. ([#1731](https://github.com/getsentry/relay/pull/1731))
+- Add count transactions toward root project. ([#1734](https://github.com/getsentry/relay/pull/1734))
+
+**Bug Fixes**:
+
+- Improve performance for replays, especially memory usage during data scrubbing.([#1800](https://github.com/getsentry/relay/pull/1800))
 
 **Internal**:
 
 - Add metric name as tag on Sentry errors from relay dropping metrics. ([#1797](https://github.com/getsentry/relay/pull/1797))
+- Make sure to scrub all the fields with PII. If the fields contain an object, the entire object will be removed. ([#1789](https://github.com/getsentry/relay/pull/1789))
+- Keep meta for removed custom measurements. ([#1815](https://github.com/getsentry/relay/pull/1815))
+- Drop replay recording payloads if they cannot be parsed or scrubbed. ([#1683](https://github.com/getsentry/relay/pull/1683))
 
 ## 23.1.1